# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Knull Incident API
  version: 0.1.0
  description: >
    HTTP contract for the Knull incident API. This foundation slice defines the
    health endpoints; incident, service, and approval operations are added by
    later implementation tasks. The TypeScript client types are generated from
    this document.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/me:
    get:
      operationId: getMe
      summary: Current operator identity
      responses:
        "200":
          description: The signed-in operator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operator"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/integrations/credentials:
    get:
      operationId: listCredentials
      summary: List integration credential metadata
      description: Returns metadata only; secret values are never included.
      responses:
        "200":
          description: Stored credential metadata.
          content:
            application/json:
              schema:
                type: object
                required: [credentials]
                properties:
                  credentials:
                    type: array
                    items:
                      $ref: "#/components/schemas/CredentialMetadata"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putCredential
      summary: Store or replace an integration credential
      description: >
        Accepts a write-only secret value, encrypts it at rest, and returns
        metadata only. The value is never returned by any endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutCredentialRequest"
      responses:
        "200":
          description: Stored credential metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialMetadata"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/services:
    get:
      operationId: listServices
      summary: List configured services
      responses:
        "200":
          description: All configured services.
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createService
      summary: Create a service mapping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Duplicate key in this environment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/services/{id}:
    get:
      operationId: getService
      summary: Get a service
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: The service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateService
      summary: Update a service mapping
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Duplicate key in this environment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/services/{id}/enabled:
    patch:
      operationId: setServiceEnabled
      summary: Enable or disable a service
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          description: Updated service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/services/{id}/recovery-policy:
    get:
      operationId: getRecoveryPolicy
      summary: Get the latest versioned recovery policy for a service
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: Current immutable policy snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryPolicySnapshot"
        "401": { description: Authentication required. }
        "404": { description: No policy configured for this service. }
    put:
      operationId: putRecoveryPolicy
      summary: Create an immutable recovery policy version
      description: Identical writes return the existing version. Workload readiness and at least one service indicator are required; missing observations can never count as recovery.
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryPolicy"
      responses:
        "200":
          description: Current immutable policy snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryPolicySnapshot"
        "400": { description: Invalid thresholds, signals, or observation window. }
        "401": { description: Authentication required. }
        "404": { description: Service not found. }
  /api/services/{id}/recovery-policy/versions/{version}:
    get:
      operationId: getRecoveryPolicyVersion
      summary: Get the exact policy used by a historic assessment
      parameters:
        - $ref: "#/components/parameters/ServiceId"
        - name: version
          in: path
          required: true
          schema: { type: integer, format: int64, minimum: 1 }
      responses:
        "200":
          description: Requested immutable policy snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryPolicySnapshot"
        "400": { description: Invalid version. }
        "401": { description: Authentication required. }
        "404": { description: Policy version not found. }
  /api/alerts/webhook:
    post:
      operationId: alertWebhook
      summary: Prometheus Alertmanager webhook
      description: >
        Receives Alertmanager notifications. Authenticated by a shared secret
        (Bearer token or `token` query param), not an operator session.
        Unauthorized or malformed deliveries are recorded and rejected.
      responses:
        "202":
          description: Delivery processed; per-alert outcomes returned.
        "400":
          description: Malformed payload.
        "401":
          description: Invalid or missing secret.
  /api/fleet:
    get:
      operationId: getFleet
      summary: Fleet health and active incidents
      responses:
        "200":
          description: Fleet rows, services with active incidents first.
          content:
            application/json:
              schema:
                type: object
                required: [fleet]
                properties:
                  fleet:
                    type: array
                    items:
                      $ref: "#/components/schemas/FleetRow"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/incidents:
    get:
      operationId: listIncidents
      summary: List incidents
      responses:
        "200":
          description: Incidents, most recently updated first.
          content:
            application/json:
              schema:
                type: object
                required: [incidents]
                properties:
                  incidents:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: startInvestigation
      summary: Start an operator-initiated investigation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [serviceId, summary]
              properties:
                serviceId:
                  type: string
                  format: uuid
                summary:
                  type: string
                symptoms:
                  type: object
                  additionalProperties:
                    type: string
                force:
                  type: boolean
      responses:
        "201":
          description: Incident opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid request or unknown service.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: An active investigation already exists for this service.
  /api/incidents/{id}:
    get:
      operationId: getIncident
      summary: Get an incident
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: The incident.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/incidents/{id}/collect:
    post:
      operationId: collectEvidence
      summary: Run read-only investigation for an incident
      description: >
        Runs the connected read-only investigators and records evidence. When
        Jev is configured, classifies incident causes from available evidence,
        records ranked hypotheses, and escalates uncertain results.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "202":
          description: Evidence collected.
        "400":
          description: Incident has no service mapping.
        "404":
          description: Incident not found.
        "503":
          description: No investigation integrations connected.
        "502":
          description: Evidence collected, but Jev classification failed.
  /api/incidents/{id}/events:
    get:
      operationId: listIncidentEvents
      summary: Get an incident's audit history
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Ordered audit events.
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/IncidentEvent"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/incidents/{id}/plans:
    post:
      operationId: createActionPlan
      summary: Record a bounded remediation proposal
      description: Seals a versioned action scoped to the incident service mapping and stored evidence. Does not execute it.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionDraft"
      responses:
        "201":
          description: Sealed action proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedActionPlan"
        "400":
          description: Invalid action or scope.
        "401":
          description: Authentication required.
        "404":
          description: Incident not found.
        "503":
          description: Action planning unavailable.
  /api/incidents/{id}/sandbox-runs:
    post:
      operationId: prepareSandboxRun
      summary: Prepare and clean up a dedicated-cluster sandbox run
      description: Creates a restricted run namespace for a sealed action and records its provenance. Prepared does not mean validation passed.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actionEventId, workload]
              properties:
                actionEventId: { type: string, format: uuid }
                workload:
                  $ref: "#/components/schemas/SandboxWorkload"
      responses:
        "202":
          description: Run prepared and cleanup verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxRecordedRun"
        "400":
          description: Invalid request.
        "401":
          description: Authentication required.
        "404":
          description: Incident not found.
        "422":
          description: Sandbox isolation, execution, or cleanup failed.
        "503":
          description: Dedicated sandbox cluster not configured.
  /api/incidents/{id}/validations/memory:
    post:
      operationId: validateMemory
      summary: Reproduce a 256Mi OOM and validate a 1Gi candidate
      description: Runs two isolated sandbox namespaces, requires a TrueForge executed script and independent Kubernetes and Prometheus observations, verifies cleanup, then records the result. Failed or missing evidence cannot advance to approval.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actionEventId, workload]
              properties:
                actionEventId: { type: string, format: uuid }
                workload: { $ref: "#/components/schemas/SandboxWorkload" }
      responses:
        "200":
          description: Baseline failure reproduced and candidate passed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MemoryValidationResult" }
        "400": { description: Invalid request. }
        "401": { description: Authentication required. }
        "404": { description: Incident not found. }
        "422":
          description: Validation failed; response includes result and error.
          content:
            application/json:
              schema:
                type: object
                required: [error, result]
                properties:
                  error: { type: string }
                  result: { $ref: "#/components/schemas/MemoryValidationResult" }
        "503": { description: Validation dependencies not configured. }
  /api/incidents/{id}/approvals:
    post:
      operationId: approveAction
      summary: Approve one validated action for 15 minutes
      description: Records the signed-in operator and exact sealed action digest. Production remains unchanged until the backend execution gate checks the live precondition.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionDecisionRequest"
      responses:
        "201":
          description: Action-specific approval recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionDecision"
        "400": { description: Invalid request. }
        "401": { description: Authentication required. }
        "404": { description: Incident not found. }
        "409": { description: Action was not eligible or was already decided. }
        "503": { description: Approval service unavailable. }
  /api/incidents/{id}/denials:
    post:
      operationId: denyAction
      summary: Deny one validated action
      description: Records the signed-in operator and optional reason, then atomically moves the incident to DENIED.
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionDecisionRequest"
      responses:
        "201":
          description: Action-specific denial recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionDecision"
        "400": { description: Invalid request. }
        "401": { description: Authentication required. }
        "404": { description: Incident not found. }
        "409": { description: Action was not eligible or was already decided. }
        "503": { description: Approval service unavailable. }
components:
  parameters:
    ServiceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IncidentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    RecoveryPolicy:
      type: object
      required: [windowSeconds, missingData, workload]
      properties:
        windowSeconds: { type: integer, minimum: 60, maximum: 3600 }
        missingData: { type: string, enum: [UNCERTAIN, NOT_RECOVERED] }
        workload:
          type: object
          required: [minReadyFraction]
          properties:
            minReadyFraction: { type: number, exclusiveMinimum: 0, maximum: 1 }
        errorRate:
          type: object
          required: [maxRatio, minSamples]
          properties:
            maxRatio: { type: number, minimum: 0, exclusiveMaximum: 1 }
            minSamples: { type: integer, minimum: 10 }
        latencyP95:
          type: object
          required: [maxP95Milliseconds, minSamples]
          properties:
            maxP95Milliseconds: { type: number, exclusiveMinimum: 0 }
            minSamples: { type: integer, minimum: 10 }
    RecoveryPolicySnapshot:
      type: object
      required: [serviceId, version, digest, policy, configuredBy, configuredAt]
      properties:
        serviceId: { type: string, format: uuid }
        version: { type: integer, format: int64 }
        digest: { type: string }
        policy: { $ref: "#/components/schemas/RecoveryPolicy" }
        configuredBy: { type: string, format: uuid }
        configuredAt: { type: string, format: date-time }
...
    Operator:
      type: object
      required: [id, email, name, createdAt, lastLogin]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
    CredentialMetadata:
      type: object
      required: [id, kind, scope, fingerprint, createdBy, createdAt, updatedAt]
      description: Metadata about a stored credential. Never contains the value.
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [kubernetes, prometheus, github, openai]
        scope:
          type: string
          enum: [read, sandbox, production]
        fingerprint:
          type: string
          description: Non-reversible identifier of the secret value.
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PutCredentialRequest:
      type: object
      required: [kind, scope, value]
      properties:
        kind:
          type: string
          enum: [kubernetes, prometheus, github, openai]
        scope:
          type: string
          enum: [read, sandbox, production]
        value:
          type: string
          description: Write-only secret value; never returned by any endpoint.
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
          description: Field name to actionable message.
    ServiceInput:
      type: object
      required: [key, displayName, environment, k8sCluster, k8sNamespace, k8sWorkload]
      properties:
        key:
          type: string
          description: Stable slug identifier (e.g. checkout-api).
        displayName:
          type: string
        environment:
          type: string
          description: Explicit environment slug (e.g. production).
        k8sCluster:
          type: string
        k8sNamespace:
          type: string
        k8sWorkload:
          type: string
        prometheusLabels:
          type: object
          additionalProperties:
            type: string
        githubRepo:
          type: string
          description: owner/repo (optional).
        githubRef:
          type: string
        recoveryPolicyRef:
          type: string
    Service:
      allOf:
        - $ref: "#/components/schemas/ServiceInput"
        - type: object
          required: [id, enabled, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            enabled:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    Incident:
      type: object
      required:
        [id, serviceKey, environment, state, version, correlationId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        serviceKey:
          type: string
        environment:
          type: string
        alertIdentity:
          type: string
        summary:
          type: string
        symptoms:
          type: object
          additionalProperties:
            type: string
        state:
          $ref: "#/components/schemas/IncidentState"
        version:
          type: integer
          format: int64
        correlationId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    FleetRow:
      type: object
      required: [key, displayName, environment, enabled, status, activeIncidentCount, signalsAvailable]
      properties:
        serviceId:
          type: string
        key:
          type: string
        displayName:
          type: string
        environment:
          type: string
        enabled:
          type: boolean
        status:
          type: string
          enum: [HEALTHY, INVESTIGATING, AWAITING_APPROVAL, REMEDIATING, VERIFYING, INCIDENT]
        activeIncidentCount:
          type: integer
        latestIncident:
          type: object
          properties:
            id:
              type: string
            state:
              $ref: "#/components/schemas/IncidentState"
            summary:
              type: string
        signalsAvailable:
          type: boolean
    IncidentEvent:
      type: object
      required: [id, incidentId, seq, type, actor, createdAt]
      properties:
        id:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        seq:
          type: integer
          format: int64
        type:
          type: string
          enum: [STATE_CHANGE, NOTE]
        category:
          type: string
          enum: [system, observation, hypothesis, decision, action]
        source:
          type: string
        target:
          type: string
        observedAt:
          type: string
          format: date-time
        fromState:
          $ref: "#/components/schemas/IncidentState"
        toState:
          $ref: "#/components/schemas/IncidentState"
        actor:
          type: string
        reason:
          type: string
        correlationId:
          type: string
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time